Remove results section
