+++ src/ensembl/src/content/app/custom-download/containers/content/preview-download/PreviewDownload.tsx
@@ -14,7 +14,7 @@ import {
   flattenObject,
   attributeDisplayNames,
   getProcessedFilters
-} from '../result-holder/resultHolderHelper';
+} from '../result-loader/resultLoaderHelper';
 
 import {
   setShowPreview,
